# openapi.json
{"openapi":"3.0.2","info":{"title":"DeepFake Detection Platform","description":"API for deepfake detection across images, audio, and video","version":"1.0.0"},"paths":{"/api/v1/upload/upload":{"post":{"tags":["upload"],"summary":"Upload File","operationId":"upload_file_api_v1_upload_upload_post","requestBody":{"content":{"multipart/form-data":{"schema":{"$ref":"#/components/schemas/Body_upload_file_api_v1_upload_upload_post"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/DetectionResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/detection/status/{task_id}":{"get":{"tags":["detection"],"summary":"Get Task Status","description":"Get the status of a detection task.","operationId":"get_task_status_api_v1_detection_status__task_id__get","parameters":[{"required":true,"schema":{"title":"Task Id","type":"string"},"name":"task_id","in":"path"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"title":"Response Get Task Status Api V1 Detection Status  Task Id  Get","type":"object"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/detection/results":{"get":{"tags":["detection"],"summary":"Get Recent Results","description":"Get recent detection results.","operationId":"get_recent_results_api_v1_detection_results_get","parameters":[{"required":false,"schema":{"title":"Limit","type":"integer","default":10},"name":"limit","in":"query"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"title":"Response Get Recent Results Api V1 Detection Results Get","type":"array","items":{"$ref":"#/components/schemas/DetectionResult"}}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/detection/results/{result_id}":{"get":{"tags":["detection"],"summary":"Get Result","description":"Get a specific detection result.","operationId":"get_result_api_v1_detection_results__result_id__get","parameters":[{"required":true,"schema":{"title":"Result Id","type":"string"},"name":"result_id","in":"path"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/DetectionResult"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/results/query":{"post":{"tags":["results"],"summary":"Query Result","description":"Query the status or result of a detection task.","operationId":"query_result_api_v1_results_query_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ResultQuery"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ResultStatus"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/results/detail/{result_id}":{"get":{"tags":["results"],"summary":"Get Detailed Result","description":"Get detailed result for a completed detection.","operationId":"get_detailed_result_api_v1_results_detail__result_id__get","parameters":[{"description":"Result ID to retrieve","required":true,"schema":{"title":"Result Id","type":"string","description":"Result ID to retrieve"},"name":"result_id","in":"path"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/DetailedResult"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/health":{"get":{"summary":"Health Check","description":"Health check endpoint for monitoring.","operationId":"health_check_health_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/":{"get":{"summary":"Root","description":"Root endpoint with API information.","operationId":"root__get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}}},"components":{"schemas":{"Body_upload_file_api_v1_upload_upload_post":{"title":"Body_upload_file_api_v1_upload_upload_post","required":["file"],"type":"object","properties":{"file":{"title":"File","type":"string","format":"binary"},"detection_params":{"$ref":"#/components/schemas/DetectionRequest"}}},"DetailedResult":{"title":"DetailedResult","required":["id","is_fake","confidence_score","media_type","detection_details","models_used"],"type":"object","properties":{"id":{"title":"Id","type":"string","description":"Unique identifier for the result","format":"uuid"},"is_fake":{"title":"Is Fake","type":"boolean","description":"Whether the media is detected as fake"},"confidence_score":{"title":"Confidence Score","type":"number","description":"Overall confidence score"},"media_type":{"title":"Media Type","type":"string","description":"Type of media that was analyzed"},"detection_details":{"title":"Detection Details","type":"object","description":"Detailed detection information"},"models_used":{"title":"Models Used","type":"object","additionalProperties":{"type":"string"},"description":"Models used for the detection"},"visualizations":{"title":"Visualizations","allOf":[{"$ref":"#/components/schemas/VisualizationData"}],"description":"Visualization data"}},"example":{"id":"abcd1234-ef56-7890-abcd-1234ef567890","is_fake":true,"confidence_score":0.92,"media_type":"image","detection_details":{"face_manipulation_score":0.95,"inconsistency_score":0.88,"artifact_score":0.91},"models_used":{"primary":"XceptionNet","secondary":"EfficientNet-B4"}}},"DetectionRequest":{"title":"DetectionRequest","required":["media_type"],"type":"object","properties":{"media_type":{"title":"Media Type","type":"string","description":"Type of media: image, audio, video"},"detailed_analysis":{"title":"Detailed Analysis","type":"boolean","description":"Whether to perform detailed analysis","default":false},"confidence_threshold":{"title":"Confidence Threshold","type":"number","description":"Confidence threshold for detection"}}},"DetectionResponse":{"title":"DetectionResponse","required":["task_id","status","media_type","estimated_time"],"type":"object","properties":{"task_id":{"title":"Task Id","type":"string","description":"Task ID for tracking the detection process"},"status":{"title":"Status","type":"string","description":"Status of the detection task"},"media_type":{"title":"Media Type","type":"string","description":"Type of media being processed"},"estimated_time":{"title":"Estimated Time","type":"integer","description":"Estimated processing time in seconds"}}},"DetectionResult":{"title":"DetectionResult","required":["id","file_hash","media_type","is_fake","confidence_score","created_at"],"type":"object","properties":{"id":{"title":"Id","type":"string","description":"Unique identifier for the detection result","format":"uuid"},"file_hash":{"title":"File Hash","type":"string","description":"Hash of the analyzed file"},"media_type":{"title":"Media Type","type":"string","description":"Type of media that was analyzed"},"is_fake":{"title":"Is Fake","type":"boolean","description":"Whether the media is detected as fake"},"confidence_score":{"title":"Confidence Score","type":"number","description":"Confidence score of the detection"},"created_at":{"title":"Created At","type":"string","description":"When the detection was performed","format":"date-time"},"detection_details":{"title":"Detection Details","type":"object","description":"Detailed detection results"},"models_used":{"title":"Models Used","type":"object","additionalProperties":{"type":"string"},"description":"Models used for detection"},"visualizations":{"title":"Visualizations","type":"object","additionalProperties":{"type":"string"},"description":"Paths to visualization files"}}},"HTTPValidationError":{"title":"HTTPValidationError","type":"object","properties":{"detail":{"title":"Detail","type":"array","items":{"$ref":"#/components/schemas/ValidationError"}}}},"HeatmapData":{"title":"HeatmapData","required":["url","width","height","regions"],"type":"object","properties":{"url":{"title":"Url","type":"string","description":"URL to the heatmap image"},"width":{"title":"Width","type":"integer","description":"Width of the image"},"height":{"title":"Height","type":"integer","description":"Height of the image"},"regions":{"title":"Regions","type":"array","items":{"type":"object"},"description":"Regions of interest with scores"}}},"ResultQuery":{"title":"ResultQuery","type":"object","properties":{"task_id":{"title":"Task Id","type":"string","description":"Task ID from detection request"},"result_id":{"title":"Result Id","type":"string","description":"Result ID from previous query","format":"uuid"}},"example":{"task_id":"abcd1234-ef56-7890-abcd-1234ef567890"}},"ResultStatus":{"title":"ResultStatus","required":["status","progress"],"type":"object","properties":{"status":{"title":"Status","type":"string","description":"Status of the detection process"},"progress":{"title":"Progress","type":"number","description":"Progress as a percentage"},"message":{"title":"Message","type":"string","description":"Additional status message"},"result_id":{"title":"Result Id","type":"string","description":"Result ID if complete","format":"uuid"}},"example":{"status":"processing","progress":65.5,"message":"Processing video frames"}},"TemporalData":{"title":"TemporalData","required":["timestamps","values","threshold"],"type":"object","properties":{"timestamps":{"title":"Timestamps","type":"array","items":{"type":"number"},"description":"Timestamps for time-series data"},"values":{"title":"Values","type":"array","items":{"type":"number"},"description":"Values for each timestamp"},"threshold":{"title":"Threshold","type":"number","description":"Threshold line value"}}},"ValidationError":{"title":"ValidationError","required":["loc","msg","type"],"type":"object","properties":{"loc":{"title":"Location","type":"array","items":{"anyOf":[{"type":"string"},{"type":"integer"}]}},"msg":{"title":"Message","type":"string"},"type":{"title":"Error Type","type":"string"}}},"VisualizationData":{"title":"VisualizationData","type":"object","properties":{"heatmap":{"title":"Heatmap","allOf":[{"$ref":"#/components/schemas/HeatmapData"}],"description":"Heatmap visualization data"},"temporal":{"title":"Temporal","allOf":[{"$ref":"#/components/schemas/TemporalData"}],"description":"Temporal analysis data"},"frequency":{"title":"Frequency","type":"object","description":"Frequency analysis data"}}}}}}